Fix connection closed before file send completed

+++ src/Connection/TcpConnection.php
@@ -723,6 +723,9 @@ class TcpConnection extends ConnectionInterface implements \JsonSerializable
                 }
             }
             if ($this->_status === self::STATUS_CLOSING) {
+                if ($this->__streamSending) {
+                    return true;
+                }
                 $this->destroy();
             }
             return true;
